backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
langchain==0.3.13
langchain-google-genai==2.0.8
PyPDF2==3.0.1
scikit-learn==1.6.1
firebase-admin==6.6.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
numpy==2.2.1
pandas==2.2.3
python-socketio==5.11.0
cloudinary==1.41.0
Pillow==11.0.0
google-generativeai==0.8.3
python-dateutil==2.9.0
email-validator>=2.0
python-docx==1.1.2
